[update] fixed folder structure to reflect miro and design decisions
