+++ MealFridge/Views/Search/Index.cshtml
@@ -5,14 +5,14 @@
             <input id="sbn" type="text" class="form-control" placeholder="Hot Chili" aria-label="Search for recipes by name">
             <div class="input-group-append">
                 <button class="btn btn-success" onclick="searchByName()" type="button">
-                <i class="fa fa-search" aria-hidden="true"></i>
-            </button>
+                    <i class="fa fa-search" aria-hidden="true"></i>
+                </button>
+            </div>
         </div>
-    </div>
     </div>
     <div id="main" class="card-columns"></div>
 </div>
 
 @section Scripts {
-    <script src="~/js/SearchByName/Index.js"></script>
+    <script src="~/js/Search/Index.js"></script>
 }
