--- MealFridge/wwwroot/js/Models/SearchByRecipe/RecipeDetails.js
@@ -1,3 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=RecipeDetails.js.map
