+++ MealFridge/wwwroot/js/Search/Index.js
